Fixed merge conflicts. Resolved double fragment and viewmodel classes. Rearranged things into correct packages. And fixed a constraint error in fragment_edit_schedule.xml
